feat!: drop unneeded create-fetchival
